Extract to common lib
