+++ truck-fleet-manager/projects/app-login/src/lib/login-form/login-form.component.ts
@@ -4,10 +4,8 @@ import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { JwtHelperService } from '@auth0/angular-jwt';
 
 import * as CryptoJS from 'crypto-js';
-import { FormsValidationMessages } from 'projects/common/src/constants';
+import { AppLoginService, FormsValidationMessages, UserRoles } from 'projects/common/src';
 
-import { AppLoginService } from '../services/app-login.service';
-import { UserRoles } from '../models/user.roles.enum';
 
 @Component({
   selector: 'fm-login-form',
