--- truck-fleet-manager/projects/app-login/src/lib/services/app-login.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { AppLoginService } from './app-login.service';
-
-describe('AppLoginService', () => {
-  let service: AppLoginService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(AppLoginService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
