+++ truck-fleet-manager/projects/common/src/constants/index.ts
@@ -1,1 +1,2 @@
 export * from './forms.validation.messages.enum';
+export * from './user.roles.enum';
