+++ truck-fleet-manager/projects/common/src/guards/admin.guard.ts
@@ -1,7 +1,8 @@
 import { CanActivateFn, Router } from '@angular/router';
-import { AppLoginService } from '../../../app-login/src/lib/services/app-login.service';
 import { inject } from '@angular/core';
-import { UserRoles } from 'projects/app-login/src/lib/models/user.roles.enum';
+
+import { AppLoginService } from '../services/app-login.service';
+import { UserRoles } from '../constants';
 
 
 export const adminGuard: CanActivateFn = (route, state) => {
