+++ truck-fleet-manager/projects/common/src/guards/operative.guard.ts
@@ -1,7 +1,8 @@
 import { inject } from '@angular/core';
 import { CanActivateFn, Router } from '@angular/router';
-import { AppLoginService } from 'projects/app-login/src/lib/services/app-login.service';
-import { UserRoles } from 'projects/app-login/src/lib/models/user.roles.enum';
+
+import { AppLoginService } from '../services';
+import { UserRoles } from '../constants';
 
 export const operativeGuard: CanActivateFn = (route, state) => {
   const loginService = inject(AppLoginService)
@@ -9,9 +10,9 @@ export const operativeGuard: CanActivateFn = (route, state) => {
   let roles: UserRoles[] = loginService.getUserRole();
 
   if (loginService.isLoggedIn() && roles.some(role => role === UserRoles.Operative)) {
-    return true; 
+    return true;
   } else {
     router.navigate(['login']);
-    return false; 
+    return false;
   }
 };
