+++ truck-fleet-manager/projects/common/src/models/index.ts
@@ -1,1 +1,2 @@
 export * from './entity.list.response';
+export * from './user';
