+++ truck-fleet-manager/projects/common/src/models/user.ts
