+++ truck-fleet-manager/projects/common/src/services/index.ts
@@ -1,1 +1,2 @@
 export * from './notification.service';
+export * from './app-login.service';
